created a global mobile menu and page background
