+++ src/app/globals.css
@@ -34,6 +34,9 @@
 :root {
   --background: #ffffff;
   --foreground: #171717;
+/* App theme colors */
+  --background-color: #1F2937;
+  --text-color: #ffffff;
 }
 
 @media (prefers-color-scheme: dark) {
@@ -44,8 +47,9 @@
 }
 
 body {
-  color: var(--foreground);
-  background: var(--background);
+  color: var(--text-color);
+    background: var(--background-color);
   font-family: Arial, Helvetica, sans-serif;
+  min-height: 100vh;
 }
